refactor: modularize stock calculation services and introduce constants
- Refactored the `Calculator` class to utilize new modular services for single stock and overall calculations, improving code organization and readability.
- Introduced a new `constants.py` file to define thresholds for price and holding counts, enhancing maintainability.
- Added `money_weighted.py` for calculating money-weighted returns, and `single_metrics.py` for computing single stock metrics in a single pass.
- Created `overall.py` to handle overall metrics and XIRR calculations, streamlining the overall data computation process.
- Updated `single_stock.py` to integrate new metrics and improve stock data assembly logic.
