+++ stockManager/backend/services/calculation/constants.py
@@ -1,0 +1,5 @@
+"""盈亏计算阈值常量"""
+
+MIN_PRICE_THRESHOLD = 0.001  # 最小价格阈值，低于此值认为价格无效
+MIN_VALUE_THRESHOLD = 0.1  # 最小市值阈值，用于判断昨日市值是否有效
+MIN_HOLD_COUNT_THRESHOLD = 0.001  # 最小持股数阈值，用于浮点数比较
